refactor: 컨트롤러 클래스 네이밍 변경 및 패키지 분리
