+++ src/main/java/roomescape/controller/view/AdminController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.view;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 
 @RequestMapping("/admin")
 @Controller
-public class RoomescapeAdminController {
+public class AdminController {
 
     @GetMapping
     public String getAdminPage() {
