+++ src/test/java/roomescape/e2e/LayeredArchitectureE2ETest.java
@@ -1,5 +1,7 @@
 package roomescape.e2e;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 import java.lang.reflect.Field;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,16 +9,14 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.controller.RoomescapeAdminController;
-
-import static org.assertj.core.api.Assertions.assertThat;
+import roomescape.controller.view.AdminController;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 public class LayeredArchitectureE2ETest {
 
     @Autowired
-    private RoomescapeAdminController roomescapeAdminController;
+    private AdminController roomescapeAdminController;
 
     @Test
     @DisplayName("컨트롤러는 jdbcTemplate를 의존하지 않아야 한다")
